Removed beenInfected, added ages, started community function
beenInfected member has been removed.

Added an age check to assign to each person, still needs to be properly calculated using the percentages.

Began creation of the community places function. This function needs to created several community places. Then, during the day-to-day kernel invocation, we need to assign these people to one of these places randomly every day at 5-7.

Also, hasInfected is not being used anywhere in the code. We need to incorporate this!