Added age group
beenInfected member has been removed.

Added an age assignment to each thread. This properly reflects age group proportional to the stats from the input file.

Began creation of the community places function. This function needs to created several community places. Then, during the day-to-day kernel invocation, we need to assign these people to one of these places randomly every day at 5-7.

Also, hasInfected is not being used anywhere in the code. We need to incorporate this!